Added maze saving/loading. Small changes in how Maze class works. Removed colors when generating maze.

+++ MainProject/fps_counter.cpp
@@ -96,7 +96,7 @@ void fps_counter::setUpdateInterval(unsigned int ns){
 	update_interval = ns;
 }
 
-float fps_counter::getFrameCount(){
+GLuint fps_counter::getFrameCount(){
 	return frames_count;
 }
